feat: switch workflow result downloads to use browser native downloads (#3728)
### Purpose ### 
This pull request partially addresses issue #3404 by switching workflow
result downloads to use the browser's native download functionality.
Using the browser to handle downloads enables the browser's built-in
download UI, including a progress bar when downloading larger files.

### Changes ### 
Previously, we used Angular's `HttpClient` to request the result file
from the backend as a `Blob` and download the `Blob` to the local
filesystem. Now, we use a standard HTML form submission to request the
result file, allowing the browser to handle the download. The backend
logic for generating the file stream remains the same.

Headers such as `Authorization` (used for Bearer tokens) cannot be added
to standard form submissions. To work around this, we pass the JWT token
directly as a form parameter and manually check for validity of the
token in the backend. We also manually verify that the user role is
valid.

#### Summary of Changes ####
- Changed frontend logic for local workflow result downloads to use a
standard form submission instead of `HttpClient` to request the
download, allowing the browser to handle the download
- Added endpoint to receive form request and authenticate the request
via the JWT token before triggering download

### Demonstration ###

https://github.com/user-attachments/assets/a2d5278e-5ebf-441f-99bf-ff93d1667271

---------

Signed-off-by: Xinyuan Lin <xinyual3@uci.edu>
Co-authored-by: Chris <143021053+kunwp1@users.noreply.github.com>
Co-authored-by: ali risheh <ali.risheh876@gmail.com>
Co-authored-by: ali <a_icam@yahoo.com>
Co-authored-by: Xinyuan Lin <xinyual3@uci.edu>