chore: relocate root project build out from core (#3852)
According to the reorganization plan #3846, we want to move components
(such as services) out from `core/` to be in root. This PR is the firs
step to do so by declaring the sbt project in root. To reduce the name
`Core`, this PR also renames the root project to `Texera`. This can also
better align with our future release name.

resolves #3854