chore: add test-ci branch protection rules (#3937)
### What changes were proposed in this PR?
This PR is the initial setup for testing #3934. We create a new branch
`test-ci` and add the branch protection rules against `test-ci` for
testing purposes.

The rules names are exact string matches from Github Actions.


### Any related issues, documentation, discussions?
related to #3934

### How was this PR tested?
N/A. It should not affect any other branches. We are manually testing.

### Was this PR authored or co-authored using generative AI tooling?
No.